+++ src/codegen/codegen_decl_defs.c
@@ -131,7 +131,7 @@ static void emit_type_aliases_internal(ParserContext *ctx, ASTNode *node, Visite
                 if (strstr(c_type_str, "(*)"))
                 {
                     char *ptr = strstr(c_type_str, "(*)");
-                    int prefix_len = ptr - c_type_str;
+                    ptrdiff_t prefix_len = ptr - c_type_str;
                     EMIT(ctx, "typedef %.*s (*%s)%s;\n", prefix_len, c_type_str,
                          node->type_alias.alias, ptr + 3);
                 }
@@ -181,7 +181,7 @@ void emit_global_aliases(ParserContext *ctx)
                 if (strstr(c_type_str, "(*)"))
                 {
                     char *ptr = strstr(c_type_str, "(*)");
-                    int prefix_len = ptr - c_type_str;
+                    ptrdiff_t prefix_len = ptr - c_type_str;
                     EMIT(ctx, "typedef %.*s (*%s)%s;\n", prefix_len, c_type_str, ta->alias,
                          ptr + 3);
                 }
