+++ src/platform/os.c
@@ -67,7 +67,7 @@ double z_get_monotonic_time(void)
 #else
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec + ts.tv_nsec / 1e9;
+    return (double)ts.tv_sec + (double)ts.tv_nsec / 1e9;
 #endif
 }
 
@@ -84,7 +84,7 @@ double z_get_time(void)
 #else
     struct timespec ts;
     clock_gettime(CLOCK_REALTIME, &ts);
-    return ts.tv_sec + ts.tv_nsec / 1e9;
+    return (double)ts.tv_sec + (double)ts.tv_nsec / 1e9;
 #endif
 }
